mini_design_project — копия/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-embed-video==1.4.8
idna==3.4
Pillow==9.4.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
